% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{as_urn}
\alias{as_urn}
\title{Converter em URN}
\usage{
as_urn(string)
}
\arguments{
\item{string}{Texto a ser convertido em URN}
}
\value{
O mesmo texto como URN
}
\description{
Converter em URN
}
\examples{
as_urn("urn:lex:br:federal:lei:2011-11-18;12527")
}
