% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_normas.R
\name{criar_urn}
\alias{criar_urn}
\title{Criar URN}
\usage{
criar_urn(nome_de_norma, nivel = "federal")
}
\arguments{
\item{nome_de_norma}{Nome de uma norma}

\item{nivel}{nivel da legisla\\u00e7\\u00e3o}
}
\value{
URN da norma, no padr\\u00e3o adotado pelo LEXML
(\url{http://lexml.gov.br/})
}
\description{
Criar URN
}
\examples{
# criar_urn("Lei n\\\\u00ba 12.527, de 18 de novembro de 2011")
}
