% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilidades.R
\name{eliminar_quebras}
\alias{eliminar_quebras}
\title{Eliminar quebras de texto mal localizadas}
\usage{
eliminar_quebras(texto)
}
\arguments{
\item{texto}{Um texto}
}
\value{
O mesmo sem quebras ruins
}
\description{
Eliminar quebras de texto mal localizadas
}
