% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_normas.R
\name{nome_norma}
\alias{nome_norma}
\title{Nome da norma}
\usage{
nome_norma(texto)
}
\arguments{
\item{texto}{Texto de um norma}
}
\value{
O nome da norma
}
\description{
Nome da norma
}
\examples{
texto <- texto_norma(as_urn("urn:lex:br:federal:lei:2011-11-18;12527"))
nome_norma(texto)
}
