% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_normas.R
\name{normas_citadas}
\alias{normas_citadas}
\title{Extrair normas citadas}
\usage{
normas_citadas(texto, contemnome = TRUE)
}
\arguments{
\item{texto}{Texto de um norma}

\item{contemnome}{O nome da norma est\\u00e1 contido no texto?}
}
\value{
Um vetor com o nome das normas citadas
}
\description{
Extrair normas citadas
}
\examples{
texto <- texto_norma("Lei n\\\\u00ba 12.527, de 18 de novembro de 2011")
normas_citadas(texto)
}
