% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_normas.R
\name{texto_norma}
\alias{texto_norma}
\alias{texto_norma.character}
\alias{texto_norma.urn}
\alias{texto_norma.url}
\title{Buscar texto de norma}
\usage{
texto_norma(nome, fonte)

\method{texto_norma}{character}(nome, fonte = c("senado", "planalto",
  "camara"))

\method{texto_norma}{urn}(nome, fonte = c("senado", "planalto", "camara"))

\method{texto_norma}{url}(nome, fonte = c("senado", "planalto", "camara"))
}
\arguments{
\item{nome}{Nome da norma a ser buscada. Pode ser texto, link ou urn.}

\item{fonte}{Nome da fonte. Pode ser um entre "senado", "planalto", "camara"}
}
\value{
O texto da norma
}
\description{
Buscar texto de norma
}
