% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrair_normas.R
\name{nome_norma}
\alias{nome_norma}
\title{Nome da norma}
\usage{
nome_norma(texto)
}
\arguments{
\item{texto}{Texto de um norma}
}
\value{
O nome da norma
}
\description{
Nome da norma
}
\examples{
texto <- texto_norma("Lei n\\\\u00ba 12.527, de 18 de novembro de 2011")
nome_norma(texto)
}
